tests/inputdevicecapabilities-touch.html

<!DOCTYPE html>
<title>InputDeviceCapabilities test for touch-triggered events</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevicecapabilities-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 10px;
}
button {
  margin-bottom: 20px;
}
</style>

<h1>InputDeviceCapabilities - test for touch-triggered events</h1>
<p>Tap and then drag on the touchscreen in the box below, then tap on the button.</p>
<div id="target" tabindex=1></div>
<button>Done</button>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceCapabilitiesValidator(sourceCapabilities) {
  assert_true(sourceCapabilities != null, "sourceCapabilities is set");
  assert_true(sourceCapabilities.firesTouchEvents, "sourceCapabilities.firesTouchEvents");
}

function runTests() {
  var events = ["touchstart", "touchmove", "touchend", "mousedown", "mousemove", "mouseup", "mouseenter", "mouseleave", "mouseover", "mouseout", "click", "focus", "blur"];

  // Mozilla doesn't support focusin/focusout: https://bugzilla.mozilla.org/show_bug.cgi?id=687787
  // There doesn't appear to be any good way to feature-detect for it.
  if (navigator.userAgent.indexOf("Firefox") == -1)
    events = events.concat(["focusin", "focusout"]);

  if ('PointerEvent' in window)
    events = events.concat(["pointerdown", "pointermove", "pointerup", "pointerenter", "pointerleave", "pointerover", "pointerout", "pointercancel"]);

  runInputDeviceCapabilitiesEventTests(events, inputDeviceCapabilitiesValidator);
}
</script>
